reshuffled files and namespaces
