+++ src/System/Net/Benchmarks/Apps/Tls/Quic/Server/QuicServer.csproj
@@ -17,16 +17,24 @@
     <Compile Include="QuicServerListener.cs" />
     <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkApp.cs"
              Link="Common/BenchmarkApp.cs" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkOptionsBinder.cs"
+             Link="Common/BenchmarkOptionsBinder.cs" />
     <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkServer.cs"
              Link="Common/BenchmarkServer.cs" />
     <Compile Include="$(SystemNetBenchmarksCommonPath)LogHelper.cs"
              Link="Common/LogHelper.cs" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/ApplicationProtocolConstants.cs"
+             Link="Common/Tls/ApplicationProtocolConstants.cs" />
     <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkServer.cs"
              Link="Common/Tls/TlsBenchmarkServer.cs" />
     <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkServerOptions.cs"
              Link="Common/Tls/TlsBenchmarkServerOptions.cs" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkServerOptionsBinder.cs"
+             Link="Common/Tls/TlsBenchmarkServerOptionsBinder.cs" />
     <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkOptions.cs"
              Link="Common/Tls/TlsBenchmarkOptions.cs" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkOptionsHelper.cs"
+             Link="Common/Tls/TlsBenchmarkOptionsHelper.cs" />
   </ItemGroup>
 
   <ItemGroup>
