+++ src/System/Net/Benchmarks/Common/BenchmarkApp.cs
@@ -2,31 +2,11 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.CommandLine;
-using System.CommandLine.Binding;
 using System.CommandLine.Parsing;
 
 namespace System.Net.Benchmarks;
 
-public interface IBenchmarkOptions
-{
-}
-
-public abstract class BenchmarkOptionsBinder<TOptions> : BinderBase<TOptions>
-    where TOptions : IBenchmarkOptions, new()
-{
-    public abstract void AddCommandLineArguments(RootCommand command);
-
-    protected abstract void BindOptions(TOptions options, ParseResult parsed);
-
-    protected override TOptions GetBoundValue(BindingContext bindingContext)
-    {
-        var options = new TOptions();
-        BindOptions(options, bindingContext.ParseResult);
-        return options;
-    }
-}
-
-public abstract class BenchmarkApp<TOptions> where TOptions : IBenchmarkOptions, new()
+public abstract class BenchmarkApp<TOptions> where TOptions : new()
 {
     private static bool s_appStarted;
